refactor: Create IdentityService.Contracts project for centralized contracts
