+++ services/IdentityService/src/IdentityService.API/IdentityService.API.csproj
@@ -23,6 +23,7 @@
     <ProjectReference Include="../IdentityService.Application/IdentityService.Application.csproj" />
     <ProjectReference
       Include="../IdentityService.Infrastructure/IdentityService.Infrastructure.csproj" />
+    <ProjectReference Include="../IdentityService.Contracts/IdentityService.Contracts.csproj" />
   </ItemGroup>
 
 </Project>
