+++ services/IdentityService/src/IdentityService.Application/IdentityService.Application.csproj
@@ -16,6 +16,7 @@
 
   <ItemGroup>
     <ProjectReference Include="../IdentityService.Domain/IdentityService.Domain.csproj" />
+    <ProjectReference Include="../IdentityService.Contracts/IdentityService.Contracts.csproj" />
   </ItemGroup>
 
 </Project>
