+++ services/IdentityService/src/IdentityService.Contracts/API/Auth/Requests/ConfirmEmailRequest.cs
@@ -1,10 +1,13 @@
 using System.ComponentModel.DataAnnotations;
 
-namespace IdentityService.API.Contracts.Auth;
+namespace IdentityService.Contracts.API.Auth.Requests;
 
-public class ResendEmailConfirmationRequest
+public class ConfirmEmailRequest
 {
     [Required(ErrorMessage = "Email is required")]
     [EmailAddress(ErrorMessage = "Invalid email address")]
     public required string Email { get; init; }
+
+    [Required(ErrorMessage = "Token is required")]
+    public required string Token { get; init; }
 }
