+++ services/IdentityService/src/IdentityService.Contracts/API/Auth/Requests/ResendEmailConfirmationRequest.cs
@@ -1,13 +1,10 @@
 using System.ComponentModel.DataAnnotations;
 
-namespace IdentityService.API.Contracts.Auth;
+namespace IdentityService.Contracts.API.Auth.Requests;
 
-public class ConfirmEmailRequest
+public class ResendEmailConfirmationRequest
 {
     [Required(ErrorMessage = "Email is required")]
     [EmailAddress(ErrorMessage = "Invalid email address")]
     public required string Email { get; init; }
-
-    [Required(ErrorMessage = "Token is required")]
-    public required string Token { get; init; }
 }
