+++ services/IdentityService/src/IdentityService.Contracts/Events/Users/UserCreatedEvent.cs
@@ -1,0 +1,8 @@
+namespace IdentityService.Contracts.Events.Users;
+
+public record UserCreatedEvent(
+    string UserId,
+    string Email,
+    string Username,
+    DateTime CreatedAt
+);
